requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mwapi==0.5.1
mwoauth==0.3.3
mwparserfromhell==0.5.3
oauthlib==3.0.1
PyJWT==1.7.1
PyMySQL==0.9.3
pywikibot
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
toolforge==4.2.0
urllib3==1.24.1
Werkzeug==0.15.1
